Add files via upload
I am uploading two of my attempts at the wzip. It was hard to troubleshoot for the two of these because I couldn't get past the first test, and ran the other ones individually and cross compared. The one I am submitting is the one titled "wzip.c", not my beta version, although I worked on my beta version for far longer.