Add files via upload
Initial commit for testing for Project 3. Includes the server code, dictionary used, and makefile. The code for the server is now ready for testing. In PuTTY, connection are not allowing me to use the loopback address to connect to my client. I have separately tested the check in dictionary function to ensure that it works, along with the initialization and creating of threads. Connection seems to work fine and testing is printing out all proper feedback, but when a client tries to connect connection fails.