Update server.c
I fixed the dictionary checking function so that it checks for words with a return character at the end, because that is how they are stored in 
the dictionary file. The word that is printed to the user and appended to the log file does not have the return character. I also successfully was able to test, and upon testing added many testing print statements which I plan to change so they will only run in test mode. I had connection errors originally that I was able to fix by using a different program.